The locations.json file has been updated. Now each location has a parent_id and all possible parents of known locations are in the json file. When running Carmen without allowing new locations, all the generated locations will be known (fixed ids).
Added a new option:
use_known_parent_for_unknown_places: if use_unknown_places is false,
then this option will try to find a known parent for a location. If a
known parent is found, and no location is found using another method,
the parent is used. 

Using this option will still ensure that no unknown location is
produced.